make ISerializer public; added Handles()
