--- Manatee.Json/Serialization/Internal/ISerializer.cs
@@ -1,10 +1,0 @@
-﻿namespace Manatee.Json.Serialization.Internal
-{
-	internal interface ISerializer
-	{
-		bool ShouldMaintainReferences { get; }
-
-		JsonValue Serialize<T>(T obj, JsonSerializer serializer);
-		T Deserialize<T>(JsonValue json, JsonSerializer serializer);
-	}
-}
