+++ Manatee.Json/Serialization/Internal/Serializers/BooleanSerializer.cs
@@ -1,9 +1,15 @@
-﻿namespace Manatee.Json.Serialization.Internal.Serializers
+﻿using System;
+
+namespace Manatee.Json.Serialization.Internal.Serializers
 {
 	internal class BooleanSerializer : ISerializer
 	{
 		public bool ShouldMaintainReferences => false;
 
+		public bool Handles(Type type, JsonSerializerOptions options)
+		{
+			return type == typeof(bool);
+		}
 		public JsonValue Serialize<T>(T obj, JsonSerializer serializer)
 		{
 			return (bool) (object) obj;
