+++ Manatee.Json/Serialization/Internal/Serializers/EnumNameSerializer.cs
@@ -18,6 +18,10 @@ namespace Manatee.Json.Serialization.Internal.Serializers
 
 		public bool ShouldMaintainReferences => false;
 
+		public bool Handles(Type type, JsonSerializerOptions options)
+		{
+			return type.GetTypeInfo().IsEnum && options.EnumSerializationFormat == EnumSerializationFormat.AsName;
+		}
 		public JsonValue Serialize<T>(T obj, JsonSerializer serializer)
 		{
 			var type = _GetType<T>();
