+++ Manatee.Json/Serialization/Internal/Serializers/EnumValueSerializer.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Reflection;
 
 namespace Manatee.Json.Serialization.Internal.Serializers
 {
@@ -6,6 +7,10 @@ namespace Manatee.Json.Serialization.Internal.Serializers
 	{
 		public bool ShouldMaintainReferences => false;
 
+		public bool Handles(Type type, JsonSerializerOptions options)
+		{
+			return type.GetTypeInfo().IsEnum && options.EnumSerializationFormat == EnumSerializationFormat.AsInteger;
+		}
 		public JsonValue Serialize<T>(T obj, JsonSerializer serializer)
 		{
 			var value = Convert.ToInt32(obj);
