+++ Manatee.Json/Serialization/Internal/Serializers/JsonSerializableSerializer.cs
@@ -1,9 +1,16 @@
-﻿namespace Manatee.Json.Serialization.Internal.Serializers
+﻿using System;
+using System.Reflection;
+
+namespace Manatee.Json.Serialization.Internal.Serializers
 {
 	internal class JsonSerializableSerializer : ISerializer
 	{
 		public bool ShouldMaintainReferences => true;
 
+		public bool Handles(Type type, JsonSerializerOptions options)
+		{
+			return typeof(IJsonSerializable).GetTypeInfo().IsAssignableFrom(type.GetTypeInfo());
+		}
 		public JsonValue Serialize<T>(T obj, JsonSerializer serializer)
 		{
 			var serializable = (IJsonSerializable) obj;
