+++ Manatee.Json/Serialization/Internal/Serializers/NumericSerializer.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using Manatee.Json.Internal;
 
 namespace Manatee.Json.Serialization.Internal.Serializers
 {
@@ -6,6 +7,10 @@ namespace Manatee.Json.Serialization.Internal.Serializers
 	{
 		public bool ShouldMaintainReferences => false;
 
+		public bool Handles(Type type, JsonSerializerOptions options)
+		{
+			return type.IsNumber();
+		}
 		public JsonValue Serialize<T>(T obj, JsonSerializer serializer)
 		{
 			var value = Convert.ToDouble(obj);
