+++ Manatee.Json/Serialization/Internal/Serializers/StringSerializer.cs
@@ -1,9 +1,15 @@
-﻿namespace Manatee.Json.Serialization.Internal.Serializers
+﻿using System;
+
+namespace Manatee.Json.Serialization.Internal.Serializers
 {
 	internal class StringSerializer : ISerializer
 	{
 		public bool ShouldMaintainReferences => false;
 
+		public bool Handles(Type type, JsonSerializerOptions options)
+		{
+			return type == typeof(string);
+		}
 		public JsonValue Serialize<T>(T obj, JsonSerializer serializer)
 		{
 			return obj as string;
